MSVC: ensure the macros min/max don't break Vc
windows.h defines min and max as macros. This is obviously a major bug
resulting in lots of fallout in Vc. This commit ensures that all Vc unit
tests are compiled with the presence of those evil macros, to find
issues early.
Also resolve a few missing min/max un-macroizations.

Fixes: gh-145
Signed-off-by: Matthias Kretz <kretz@kde.org>